edit-all-repositorys-noPrint

+++ src/entities/service/CartaoService.java
@@ -45,76 +45,22 @@ public class CartaoService extends Service {
         }
     }
 
-    public void cadastrarCartao(Conta conta) {
-        List<Cartao> cartoes = this.returnCartoes(conta);
-        if (cartoes != null && cartoes.size() == 2) {
-            System.err.println("Você não pode ADICIONAR mais CARTÕES, só é possível ter no MÁXIMO 2 CARTÕES");
-        } else {
-            int tipoCartao;
-
-            tipoCartao = askInt("Insira o tipo do cartão:\n[1] CRÉDITO\n[2] DÉBITO\n[3] CANCELAR");
-
-            if (tipoCartao < 1 || tipoCartao >= 3) {
-                System.out.println("Operação cancelada!");
-            } else {
-                switch (tipoCartao) {
-                    case 1 -> {
-                        CartaoDeCredito cartaoDeCredito = new CartaoDeCredito();
-                        cartaoDeCredito.setConta(conta);
-                        cartaoDeCredito.setLimite(1000);
-                        cartaoDeCredito.setTipo(TipoCartao.CREDITO);
-                        cartaoDeCredito.setVencimento(LocalDate.now().plusYears(4));
-                        cartaoDeCredito.setCodigoSeguranca(ThreadLocalRandom.current().nextInt(100, 999));
-                        cartaoDeCredito.setDataExpedicao(LocalDate.now());
-                        try {
-                            cartaoDeCredito = (CartaoDeCredito) this.cartaoRepository.adicionar(cartaoDeCredito);
-                            if (cartaoDeCredito != null) {
-                                System.out.println("Novo CARTÃO de CRÉDITO adicionado com sucesso!");
-                                System.out.println("\tDados do CARTÃO: ");
-                                System.out.println("\t\tNúmero da conta que possui o cartão: " + cartaoDeCredito.getConta().getNumeroConta());
-                                System.out.println("\t\tLimite: " + cartaoDeCredito.getLimite());
-                                System.out.println("\t\tTipo: " + cartaoDeCredito.getTipo());
-                                System.out.println("\t\tVencimento: " + cartaoDeCredito.getVencimento());
-                                System.out.println("\t\tData de expedição: " + cartaoDeCredito.getDataExpedicao());
-                                System.out.println("\t\tCódigo de segurança: " + cartaoDeCredito.getCodigoSeguranca());
-                            }
-                        } catch (BancoDeDadosException e) {
-                            e.printStackTrace();
-                        }
-                    }
-                    case 2 -> {
-                        CartaoDeDebito cartaoDeDebito = new CartaoDeDebito();
-                        cartaoDeDebito.setConta(conta);
-                        cartaoDeDebito.setTipo(TipoCartao.DEBITO);
-                        cartaoDeDebito.setCodigoSeguranca(ThreadLocalRandom.current().nextInt(100, 999));
-                        cartaoDeDebito.setDataExpedicao(LocalDate.now());
-                        cartaoDeDebito.setVencimento(LocalDate.now().plusYears(4));
-                        try {
-                            cartaoDeDebito = (CartaoDeDebito) this.cartaoRepository.adicionar(cartaoDeDebito);
-                            if (cartaoDeDebito != null) {
-                                System.out.println("Novo CARTÃO de DÉBITO adicionado com sucesso!");
-                                System.out.println("\tDados do CARTÃO: ");
-                                System.out.println("\t\tNúmero da conta que possui o cartão: " + cartaoDeDebito.getConta().getNumeroConta());
-                                System.out.println("\t\tTipo: " + cartaoDeDebito.getTipo());
-                                System.out.println("\t\tVencimento: " + cartaoDeDebito.getVencimento());
-                                System.out.println("\t\tData de expedição: " + cartaoDeDebito.getDataExpedicao());
-                                System.out.println("\t\tCódigo de segurança: " + cartaoDeDebito.getCodigoSeguranca());
-                            }
-                        } catch (BancoDeDadosException e) {
-                            e.printStackTrace();
-                        }
-                    }
-                    default -> System.err.println("Erro bizarro!");
-                }
-            }
-        }
-    }
-
     public void cadastrarCartao(Conta conta, TipoCartao tipoCartao) {
         List<Cartao> cartoes = this.returnCartoes(conta);
         if (cartoes != null && cartoes.size() == 2) {
             System.err.println("Você não pode ADICIONAR mais CARTÕES, só é possível ter no MÁXIMO 2 CARTÕES");
+        }
+
+        if(tipoCartao == null){
+
+            tipoCartao = TipoCartao.getTipoCartao(askInt("Insira o tipo do cartão:\n[1] DÉBITO\n[2] CRÉDITO\n[3] CANCELAR"))  ;
+            System.out.println(tipoCartao.getTipo());
+            if (tipoCartao.getTipo() < 1 || tipoCartao.getTipo() >= 3) {
+                System.out.println("Operação cancelada!");
+            }
         } else {
+            tipoCartao = TipoCartao.DEBITO;
+        }
             switch (tipoCartao) {
                 case CREDITO -> {
                     CartaoDeCredito cartaoDeCredito = new CartaoDeCredito();
@@ -164,7 +110,6 @@ public class CartaoService extends Service {
                 }
                 default -> System.err.println("Erro bizarro!");
             }
-        }
     }
 
     public void deletarCartao(Conta conta) {
@@ -180,15 +125,16 @@ public class CartaoService extends Service {
                     message.append("Cartão [").append(i + 1).append("] -> ").append(cartoes.get(i).getTipo() == TipoCartao.DEBITO ? "Débito" : "Crédito").append("\n");
                 }
             }
+            System.out.print("Escolha:");
             cartao = askInt(String.valueOf(message)) - 1;
-            if (cartao != -1) {
+            if (cartao > -1 && cartao < cartoes.size()) {
                 try {
                     if (cartoes.get(cartao).getTipo() == TipoCartao.CREDITO) {
                         CartaoDeCredito cartaoDeCredito = (CartaoDeCredito) cartoes.get(cartao);
-                        if (cartaoDeCredito.getLimite() != 1000 && this.cartaoRepository.remover(cartoes.get(cartao).getNumeroCartao())) {
+                        if (cartaoDeCredito.getLimite() == 1000 && this.cartaoRepository.remover(cartoes.get(cartao).getNumeroCartao())) {
                             System.out.println("CARTÃO removido com sucesso!");
                         } else {
-                            System.err.println("Problemas na deleção do CARTÃO");
+                            System.err.println("Não é possível deletar o cartão!");
                         }
                     } else {
                         if (this.cartaoRepository.remover(cartoes.get(cartao).getNumeroCartao())) {
